test: move Voodoo tests into dedicated folder
